WIP: Refactor scenes and scene controllers
