+++ karte_flip_client/src/controllers/match_scene_controllers/VsPlayerMatchSceneController.cs
@@ -1,25 +1,15 @@
-using System.Threading.Tasks;
 using Godot;
+using KarteFlipClient;
+using System.Threading.Tasks;
 
-public partial class MainVsComputerSceneController: MainSceneController {
+public partial class VsPlayerMatchSceneController: MatchSceneController{
+    private PlayerTurnManager _playerTurnManager;
 
-    ComputerTurnManager _computerTurnManager;
-
-     public override void _Ready(){
+    public override void _Ready(){
 		InitAutoLoads();
 		InitUiBindings();
 		InitializeListeners();
 		DisplayRandomCard();
-		_computerTurnManager.ComputerFirstTurn();
-	}
-
-    private void InitAutoLoads(){
-		_routeManager = GetNode<RouteManager>(
-			RouteManager.GetSingletonAutoLoad(SingletonAutoLoadEnum.ROUTE_MANAGER)
-		);
-		_computerTurnManager = GetNode<ComputerTurnManager>(
-			RouteManager.GetSingletonAutoLoad(SingletonAutoLoadEnum.COMPUTER_TURN_MANAGER)
-		);
 	}
 
     public override void _Input(InputEvent @event){
@@ -32,11 +22,20 @@ public partial class MainVsComputerSceneController: MainSceneController {
 		}
 
 		if(@event is InputEventMouseButton){
-			if(_computerTurnManager.IsMyTurn()){
+			if(_playerTurnManager.IsMyTurn()){
 				OnTapGroundTile(@event);
 			}
 		}
     }
+
+	private void InitAutoLoads(){
+		_routeManager = GetNode<RouteManager>(
+			RouteManager.GetSingletonAutoLoad(SingletonAutoLoadEnum.ROUTE_MANAGER)
+		);
+		_playerTurnManager = GetNode<PlayerTurnManager>(
+			RouteManager.GetSingletonAutoLoad(SingletonAutoLoadEnum.PLAYER_TURN_MANAGER)
+		);
+	}
 
 	private void OnTapGroundTile(InputEvent @event){
 
@@ -62,23 +61,24 @@ public partial class MainVsComputerSceneController: MainSceneController {
 			return;
 		}
 
-		_computerTurnManager.AddTokenToTilemap(tilePostion, _randomCard);
-		_tilemap.PlayTileDropAudio();
+		_playerTurnManager.ExecuteAddTokenToTilemap(tilePostion, _randomCard);
 		DisplayRandomCard();
-        _computerTurnManager.ComputerTapGroundTile();
+	}
+
+	protected override void OnPressedRestartButton(){
+		_displayDialog.ShowDialog("You can't restart a P2P match.");
+		_displayDialog.GetCancelButton().Text = "Close";
+		_displayDialog.GetConfirmButton().Text = "Ok";
+		_currentDialogue = Dialogs.RESTART_DIALOGUE;
 	}
 
 	protected override async void OnPressedDialogConfirmButton(){
 		if(Dialogs.RESTART_DIALOGUE == _currentDialogue){
-            _computerTurnManager.ResettingTurn();
-            _randomCard = TileHelper.GetRandomCard();
-			_computerTurnManager.ComputerFirstTurn();
-			_displayDialog.CloseDialog();
+			_displayDialog.CloseDialog();	
 		}
 
 		if(Dialogs.QUIT_DIALOGUE == _currentDialogue){
-			Multiplayer.MultiplayerPeer = null;
-            _computerTurnManager.ResettingTurn();
+			_playerTurnManager.ExecuteQuitMatch();
 			await Task.Delay(500);
 			_routeManager.MoveToScene(SceneFilenameEnum.LOBBY_SCENE, "Leaving game, please wait.");
 		}
