+++ karte_flip_client/src/enums/routes_enums/SceneFileNameEnum.cs
@@ -1,6 +1,6 @@
 public enum SceneFilenameEnum{
-    MAIN_VS_PLAYER_SCENE,
-    MAIN_VS_COMPUTER_SCENE,
+    VS_PLAYER_MATCH_SCENE,
+    VS_COMPUTER_MATCH_SCENE,
     LOBBY_SCENE,
     SPLASH_SCENE,
     CREDIT_SCENE,
